+++ src/index.ts
@@ -1,13 +1,14 @@
 export { DebugWeb } from './debugWeb';
+export { getLimitedArray } from './utils';
 export { createDebug } from './createDebug';
 export { getStyle, stylizeMsg } from './stylize';
 export type {
   DebugWebData,
-  DebugWebLogLevel,
   DebugWebOptions,
-  DebugWebStyle,
+  DebugWebLogLevel,
   DebugWebAliasMap,
   CreateDebugOptions,
   CustomLogLevels,
+  DebugWebStyle,
   ConsoleMethod
 } from './types';